// frontend/src/app/models/user.ts
export class User {
    id: number;
    username: string;
    password: string;
    email: string;
    firstName: string;
    lastName: string;
    birthDate: string;
    city: number;
    phoneNumber: string;
    agency: string;
    licenseNo: number;
    profileImage: string;
    userType: number;
    waitingApproval: boolean;
    denied: boolean;
}